% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{clustering_workflow}
\alias{clustering_workflow}
\title{Clustering Workflow}
\usage{
clustering_workflow(
  seu,
  excluded_cells,
  resolution = seq(0.2, 2, by = 0.2),
  organism = "human",
  experiment_name = "default_experiment",
  ...
)
}
\arguments{
\item{seu}{seurat object}

\item{excluded_cells}{named list of cells to exclude}

\item{resolution}{resolution(s) to use for clustering cells}

\item{organism}{Organism}

\item{...}{}
}
\description{
Cluster and Reduce Dimensions of a seurat object
}
\examples{
clustered_human_seu <- clustering_workflow(panc8, experiment_name = "seurat_pancreas", organism = "human")
clustered_mouse_seu <- clustering_workflow(baron2016singlecell)
}
