% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{save_seurat}
\alias{save_seurat}
\title{Save seurat object to \if{html}{\out{<project>}}/output/sce/\if{html}{\out{<feature>}}_seu.rds}
\usage{
save_seurat(seu, prefix = "unfiltered", proj_dir = getwd())
}
\arguments{
\item{proj_dir}{}

\item{...}{named arguments specifying seurat objects list of seurat objects; default "gene" and "transcript"}
}
\description{
Save seurat object to \if{html}{\out{<project>}}/output/sce/\if{html}{\out{<feature>}}_seu.rds
}
\examples{
\dontrun{
save_seurat(seu = feature_seus, proj_dir = proj_dir)

save_seurat(seu = feature_seus, prefix = "remove_nonPRs", proj_dir = proj_dir)
}
}
